correctly mount the gitserver repo mirroring disk
Prior to this change, the gitserver disk was not properly mounted which meant
that on restart gitserver would have to clone the repository in full from the
code host again, rather than using its on-disk mirror which is much faster.

This was an oversight, but easily fixed by just properly mounting the disk.