properly connect frontends to github-proxy
Prior to this change, a new connection between the frontend and github was introduced
but due to the environment variable `GITHUB_BASE_URL` defaulting to `http://github-proxy`
for our Kubernetes environments this was incorrect in our docker-compose and pure-docker
deployment environments.

In particular, the connection is made for campaigns and authz (i.e. `authorization` in
the github configuration on Sourcegraph).