cadvisor: add support for Amazon Linux (non-v2)
This volume mount is required for cadvisor to work correctly on Amazon Linux (non-v2)
host OSs, and likely other older Linux distros.

I have also tested on Ubuntu Xenial, and with a volume mount that does not exist,
and neither appear to cause any issues when `/cgroup` does not exist on the host OS
so this should be fine to add for all deployments.