Replace cpu limits with shares
In most systems we want to ensure each container gets a 'fair' amount
of CPU relative to the other resources given we have a dedicated
host.

This uses CPU shares to define the weight an the minimum reserved memory
for each service, but allows services to expand beyond that capacity unless
the system is constrained.