insights: set PGDATA=/var/lib/postgresql/data/pgdata
I previously thought `PGDATA=/var/lib/postgresql/data/pgdata` was the default,
but it is definitely not: `/var/lib/postgresql/data` is.

I removed this from `deploy-codeinsights-db.sh` for consistency with the Docker
Compose deployment, but this would have broken existing pure-docker deployments
because they were storing data in the `pgdata` subdirectory previously.

I have added this to Docker Compose deployments here, for consistency, although
it is not strictly required (Docker Compose volumes, unlike Kubernetes ones, do
not appear ot have a `lost+found` file in them and as such Postgres is happy to
use the root of the volume - whereas in Kubernetes deployments it will complain
the directory is not empty and fail to start.)

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>