Fix Jaeger ports for HTTP tracing (#1128)
Updating Docker repo to match the fix in the Helm repo from a few months
ago, so that otel-collector is sending trace data to Jaeger on the same
port that Jaeger is listening on.

### Checklist

<!--
Kubernetes and Docker Compose MUST be kept in sync. You should not merge
a change here
without a corresponding change in the other repository, unless it truly
is specific to
this repository. If uneeded, add link or explanation of why it is not
needed here.
-->
* [ ] Sister
[deploy-sourcegraph](https://github.com/sourcegraph/deploy-sourcegraph)
change:
* [ ] Sister
[customer-replica](https://github.com/sourcegraph/deploy-sourcegraph-docker-customer-replica-1)
change (if necessary, for any changes affecting pure-docker or
configuration):
* [ ] All images have a valid tag and SHA256 sum
### Test plan

Tested manually

`docker-compose -f docker-compose.yaml -f jaeger/docker-compose.yaml -f
override.yaml up -d --remove-orphans`