Close Grafana port from outside access for Docker deployments (#1204)
<!-- description here -->

Our Docker deployments have exposed the Grafana container's listening port (3370) for side-channel access by default, since we started bundling Grafana in our deployments. This hasn't been a security concern in the past because Grafana had only been used for system health and performance metrics, but Grafana will now also show PII for usage data for a select few customers adding the the Airgapped Analytics dashboard. 

Closing port 3370 on the Docker host still allows Site Admins to reach Grafana through the site admin web UI.

Also, fixed a broken port number in sourcegraph-frontend-internal's
config for the Grafana URL, which has been broken since
https://github.com/sourcegraph/deploy-sourcegraph-docker/commit/f6f8d8dff29335a4a96133b58edb0d931415fca9,
so probably not used.

### Checklist

<!--
Kubernetes and Docker Compose MUST be kept in sync. You should not merge
a change here
without a corresponding change in the other repository, unless it truly
is specific to
this repository. If unneeded, add link or explanation of why it is not
needed here.
-->
* [ ] Sister
[deploy-sourcegraph](https://github.com/sourcegraph/deploy-sourcegraph)
change:
* [ ] Sister
[customer-replica](https://github.com/sourcegraph/deploy-sourcegraph-docker-customer-replica-1)
change (if necessary, for any changes affecting pure-docker or
configuration):
* [ ] All images have a valid tag and SHA256 sum
### Test plan

Tested on test instance, frontend still proxies the connection to Grafana as needed, without side channel access

<!--
As part of SOC2/GN-104 and SOC2/GN-105 requirements, all pull requests
are REQUIRED to
provide a "test plan". A test plan is a loose explanation of what you
have done or
implemented to test this, as outlined in our Testing principles and
guidelines:

https://docs.sourcegraph.com/dev/background-information/testing_principles
  Write your test plan here after the "Test plan" header.
-->