Remove symbols statefulset (#1205)
Starting in Sourcegraph 6.4, the Symbols and Searcher services are now
combined, for easier operation, simplified architecture, and better
resource utilization.

Customers can delete the service if not deleted automatically, and may
want to consider reallocating resources previously given to symbols to
searcher. Environment variables should be copied over from symbols to
searcher.

Test plan: This deployment ran for ~2 weeks on S2 and dotcom already and
has signoff from team-product-platform.

<!-- description here -->

### Checklist

<!--
Kubernetes and Docker Compose MUST be kept in sync. You should not merge
a change here
without a corresponding change in the other repository, unless it truly
is specific to
this repository. If uneeded, add link or explanation of why it is not
needed here.
-->
* [x] Sister
[deploy-sourcegraph](https://github.com/sourcegraph/deploy-sourcegraph)
change:
* [x] Sister
[customer-replica](https://github.com/sourcegraph/deploy-sourcegraph-docker-customer-replica-1)
change (if necessary, for any changes affecting pure-docker or
configuration):
* [x] All images have a valid tag and SHA256 sum