fix(release): fix py env issue in ci-checkov (#1224)
![Screenshot 2025-06-20 at 11 56
11 AM](https://github.com/user-attachments/assets/6f27dcb0-d409-417a-a808-c0e783d879b8)
^^ we see this issue regularly during the release

It seems like a pretty generic pyenv issue.

closes REL-1044

### Test plan
Ran script locally with and without change.

<!--
As part of SOC2/GN-104 and SOC2/GN-105 requirements, all pull requests
are REQUIRED to
provide a "test plan". A test plan is a loose explanation of what you
have done or
implemented to test this, as outlined in our Testing principles and
guidelines:

https://docs.sourcegraph.com/dev/background-information/testing_principles
  Write your test plan here after the "Test plan" header.
-->