Create examples directory, and add example of external databases (#1316)
<!-- description here -->

### Checklist

<!--
Kubernetes and Docker Compose MUST be kept in sync. You should not merge
a change here
without a corresponding change in the other repository, unless it truly
is specific to
this repository. If uneeded, add link or explanation of why it is not
needed here.
-->
* [ ] Sister
[deploy-sourcegraph](https://github.com/sourcegraph/deploy-sourcegraph)
change:
* [ ] Sister
[customer-replica](https://github.com/sourcegraph/deploy-sourcegraph-docker-customer-replica-1)
change (if necessary, for any changes affecting pure-docker or
configuration):
* [ ] All images have a valid tag and SHA256 sum
### Test plan

Doc update

<!--
As part of SOC2/GN-104 and SOC2/GN-105 requirements, all pull requests
are REQUIRED to
provide a "test plan". A test plan is a loose explanation of what you
have done or
implemented to test this, as outlined in our Testing principles and
guidelines:

https://docs.sourcegraph.com/dev/background-information/testing_principles
  Write your test plan here after the "Test plan" header.
-->